requirements.txt
fastapi==0.78.0
fastapi-users[sqlalchemy]
uvicorn[standard]
alembic
aiosqlite
sqlalchemy_utils==0.38.3
asyncpg
psycopg2==2.9.3
python-dotenv
httpx==0.23.0
croniter==1.3.7
cryptography>==2.8
pyOpenSSL==22.0
ddtrace==1.4.4
